+++ remem/similarity/__init__.py
@@ -1,4 +1,11 @@
 from remem.similarity.engine import SimilarityEngine, SimilarityMatch
 from remem.similarity.index import AnnConfig
+from remem.similarity.mode import SearchMode, SearchModeResolution
 
-__all__ = ["AnnConfig", "SimilarityEngine", "SimilarityMatch"]
+__all__ = [
+    "AnnConfig",
+    "SearchMode",
+    "SearchModeResolution",
+    "SimilarityEngine",
+    "SimilarityMatch",
+]